backend/pyproject.toml
[project]
name = "radar-de-pautas"
version = "0.1.0"
description = "Radar de Pautas — Inteligência Editorial"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.0",
    # ── API ──
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    # ── Database ──
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.14.0",
    "greenlet>=3.1.0",
    # ── Messaging / Workers ──
    "celery[redis]>=5.4.0",
    "kombu>=5.4.0",
    # ── HTTP / Scraping ──
    "httpx>=0.28.0",
    "selectolax>=0.3.0",
    "trafilatura>=2.0.0",
    "playwright>=1.49.0",
    # ── PDF / OCR ──
    "pypdf>=5.0.0",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.13",
    # ── Observability ──
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "opentelemetry-instrumentation-celery>=0.50b0",
    "python-json-logger>=3.2.0",
    # ── Utils ──
    "python-dotenv>=1.0.0",
    "redis>=5.2.0",
    # ── Tests ──
    "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
